requirements.txt
Django==2.1.1
djangorestframework
gunicorn
Pillow